==> part 1.ipynb <==
# imports
import pymysql
#pymysql.install_as_MySQLdb()
from sqlalchemy import create_engine
from sqlalchemy.types import *
from sqlalchemy_utils import database_exists
import pandas as pd
import numpy as np


basics_url="https://datasets.imdbws.com/title.basics.tsv.gz"
basics = pd.read_csv(basics_url, sep='\t', low_memory=False)
#print(basics)

ratings_url="https://datasets.imdbws.com/title.ratings.tsv.gz"
ratings = pd.read_csv(ratings_url, sep='\t', low_memory=False)

akas_url="https://datasets.imdbws.com/title.akas.tsv.gz"
akas = pd.read_csv(akas_url, sep='\t', low_memory=False)

## Cleaning

basics.head()

ratings.head()

akas.head()

# Akas
     replacing \N with nan
     filtering US only movies

akas.info()

akas.replace({'\\N':np.nan}, inplace=True)

#akas = akas.loc[akas["region"] != 'US']
#akas.drop(akas.loc[akas["region"] != 'US'], inplace=True)
akas.drop(akas[akas['region'] != 'US'].index, inplace = True)

# check values in 'region'
akas['region'].value_counts()

print(akas)

# Basics

keepers =basics['tconst'].isin(akas['titleId'])
keepers

basics = basics[keepers]
basics

basics.replace({'\\N':np.nan}, inplace=True)


basics.drop(basics[basics['runtimeMinutes'].isnull()].index, inplace = True)

#import math
print(basics)
#print(math.NaN)

basics.drop(basics[basics['genres'].isnull()].index, inplace = True)

print(basics)

basics.drop(basics[basics['titleType'] != 'movie'].index, inplace = True)

print(basics)

basics['startYear'].value_counts

basics.isna().sum()
basics.info()

basics.info()

basics['startYear'] = basics['startYear'].astype('int')

# keep only 'startYear' from 2000 to 2022, including 2000 and 2022
basics = basics[(basics['startYear'] >= 2000) &
                (basics['startYear'] <= 2022)]

# check
basics['startYear'].value_counts().sort_index(ascending = False)

basics

basics.drop(basics[basics['genres'] == 'Documentary'].index, inplace = True)

basics

# Rankings

ratings.head()

ratings.info()

us_movies = basics['tconst'].isin(akas['titleId'])
us_movies
#new_ratings

basics_new = basics[us_movies]
basics_new

# create filter for ratings to only include the US by using akas
# (which has already been cut down to only include US)
us_movies = ratings['tconst'].isin(akas['titleId'])

# check
us_movies

## Review

basics_new.info()

# check info for akas
akas.info()

# check info for ratings
ratings.info()

## Save

# make data folder
import os
os.makedirs('Data/', exist_ok = True)

# check
os.listdir('Data/')

# save each file to a compressed csv file in data folder
basics_new.to_csv('Data/title_basics.csv.gz', 
              compression = 'gzip', 
              index = False)

# check
basics_new = pd.read_csv('Data/title_basics.csv.gz',
                    low_memory = False)

basics_new.head()

# save each file to a compressed csv file in data folder
akas.to_csv('Data/title_akas.csv.gz', 
              compression = 'gzip', 
              index = False)

# check
akas = pd.read_csv('Data/title_akas.csv.gz',
                    low_memory = False)

akas.head()

# save each file to a compressed csv file in data folder
ratings.to_csv('Data/title_ratings.csv.gz', 
              compression = 'gzip', 
              index = False)

# check
ratings = pd.read_csv('Data/title_ratings.csv.gz',
                    low_memory = False)

ratings.head()
