==> Part2.ipynb <==
## API Calls

* Your stakeholder wants you to extract the budget, revenue, and MPAA Rating (G/PG/PG-13/R), which is also called "Certification".

### Import libraries

# Install tmdbsimple (only need to run once)
!pip install tmdbsimple

# imports
import json, os, time
import tmdbsimple as tmdb
import pandas as pd
from tqdm.notebook import tqdm_notebook
import glob

# load api key
with open('/Users/ROSA/.secret/tmdb_api.json') as f:
    login = json.load(f)

login.keys()

import tmdbsimple as tmdb
tmdb.API_KEY =  login['api-key']

## make a movie object using the .Movies function from tmdb
movie = tmdb.Movies(603)

## movie objects have a .info dictionary 
info = movie.info()
info

info['budget']

info['revenue']

info['imdb_id']

movie = tmdb.Movies('tt1361336')
info = movie.info()
info['budget']

## Saving the Movie Certification/MPAA Rating

response = movie.releases()
response

response = movie.releases()
for c in movie.countries:
    if c['iso_3166_1'] == 'US':
        print (c['certification'])
        

# Get the movie object for the current id
movie = tmdb.Movies('tt1361336')
# save the .info .releases dictionaries
info = movie.info()
releases = movie.releases()
# Loop through countries in releases
for c in releases['countries']:
    # if the country abbreviation==US
    if c['iso_3166_1' ] =='US':
        ## save a "certification" key in the info dict with the certification
       info['certification'] = c['certification']

# function to add certification to movie.info dict
def get_movie_with_rating(movie_id):
    """Adapted from source = https://github.com/celiao/tmdbsimple"""
    
    # set movie id
    movie = tmdb.Movies(movie_id)

    # save the .info and .releases dictionaries
    info = movie.info()
    releases = movie.releases()
    
    # only get releases of the movie in the US
    for c in releases['countries']:
        
        # if country abbreviation == US
        if c['iso_3166_1'] == 'US':
            
            # save certification key into info dict
            info['certification'] = c['certification']
            
    return info

test = get_movie_with_rating("tt0848228") #put your function name here
test

## Creating folder

import json, os, time
import tmdbsimple as tmdb
from tqdm.notebook import tqdm_notebook
FOLDER = "Data/"
os.makedirs(FOLDER, exist_ok=True)
os.listdir(FOLDER)



YEARS_TO_GET = (2000,2001)

errors = [ ]

#load in the dataframe from project part 1 as basics
basics = pd.read_csv('/Users/Rosa/Documents/Data/Movie-Profitability/Data/title_basics.csv.gz')

basics.head()

def write_json(new_data,filename):
    """Adapted from source = https://geeksforgeeks.org/append-to-json-file-using-python/"""
    
    with open (filename, 'r+') as file:
        #First we load existing data into a dict.
        file_data = json.load(file)
        if (type(new_data) == list) & (type(file_data) == list) :
            file_data.extend(new_data)
        else:
            file_data.append(new_data)
        # sets file's current position at offset.
        file.seek(0)
        # convert back to json.
        json.dump(file_data,file)


# Start of OUTER loop
for YEAR in tqdm_notebook(YEARS_TO_GET, desc='YEARS', position=0):
    #Defining the JSON file to store results for year
    JSON_FILE = f'{FOLDER}tmdb_api_results_{YEAR}.json'
    # check if file exists
    file_exists = os.path.isfile(JSON_FILE)
    
    # If it does not exist: create it
    if file_exists == False:
    # save an empty dict with just "imdb_id" to the new json file.
        with open(JSON_FILE,'w') as f:
            json.dump([{'imdb_id':0}],f)

    #Saving new year as the current df
    df = basics.loc[ basics['startYear']==YEAR].copy()

    # saving movie ids to list
    movie_ids = df['tconst'].copy()
    movie_ids

    # Load existing data from json into a dataframe called "previous_df"
    previous_df = pd.read_json(JSON_FILE)
    previous_df

    # filter out any ids that are already in the JSON_FILE
    movie_ids_to_get = movie_ids[~movie_ids.isin(previous_df['imdb_id'])]

    #Get index and movie id from list
    # INNER Loop
    
    for movie_id in tqdm_notebook(movie_ids_to_get,
                                  desc=f'Movies from {YEAR}',
                                  position=1,
                                  leave=True):
        try:
            # Retrieve then data for the movie id
            temp = get_movie_with_rating(movie_id)  
            # Append/extend results to existing file using a pre-made function
            write_json(temp,JSON_FILE)
            # Short 20 ms sleep to prevent overwhelming server
            time.sleep(0.02)
            
         # if it fails, make a dict with just the id and None for certification.   
        except Exception as e:
            continue 

    final_year_df = pd.read_json(JSON_FILE)
    final_year_df.to_csv(f"{FOLDER}final_tmdb_data_{YEAR}.csv.gz", compression="gzip", index=False)

final_year_df
