==> part2_EDA.ipynb <==
## Exploratory Data Analysis
* results for each year extracted 2000 and 2001

# imports
import pandas as pd

# load data from 2000
movies_2000 = pd.read_csv('Data/final_tmdb_data_2000.csv.gz')

# check
movies_2000.info()
movies_2000.head()

# drop first row with no data
movies_2000.drop(0, inplace = True)

# check
movies_2000.head()

# load data from 2001
movies_2001 = pd.read_csv('Data/final_tmdb_data_2001.csv.gz')

# check
movies_2001.info()
movies_2001.head()

# drop first row with no data
movies_2001.drop(0, inplace = True)

# check
movies_2001.head()

# concatenate dataframes
df = pd.concat([movies_2000, movies_2001], ignore_index = True)

# check
df.info()

# save merged df to file
df.to_csv(f"Data/tmdb_results_combined.csv.gz",
          compression = "gzip",
          index = False)

# How many movies had at least some valid financial information (values > 0 for budget OR revenue)?
* Please exclude any movies with 0's for budget AND revenue from the remaining visualizations.

df = df[(df['budget'] > 0) | (df['revenue'] > 0)]
print(f"{len(df)} movies had at least some valid financial information.")

### How many movies are there in each of the certification categories (G/PG/PG-13/R)?

# use value_counts to count number of movies in each certification category
categories = df.certification.value_counts()

# loop through series to print
for index, value in categories.items():
    
    # restrict to categories asked for
    if index in ['G', 'PG', 'PG-13', 'R']:
        print(f"There are {value} movies rated {index}.")

### What is the average revenue per certification category?

# Load the data
import locale
locale.setlocale(locale.LC_ALL, '') # to put commas in large numbers

# Get the average revenue per certification category
average_revenue_per_certification_category = df.groupby('certification')['revenue'].mean()

# loop through series
for index, value in average_revenue_per_certification_category.items():
    # restrict to categories asked for
    if index in ['G', 'PG', 'PG-13', 'R']:
        
        # format numbers: 2 decimal places and commas
        formatted_value = locale.format_string('%.2f', value, grouping = True)
        # Print the results
        print(f" The average revenue of {index}-rated movies is ${formatted_value}.")

### What is the average budget per certification category?


# Get the average revenue per certification category
average_budget_per_certification_category = df.groupby('certification')['budget'].mean()

# loop through series
for index, value in average_budget_per_certification_category.items():
    # restrict to categories asked for
    if index in ['G', 'PG', 'PG-13', 'R']:
        
        # format numbers: 2 decimal places and commas
        formatted_value = locale.format_string('%.2f', value, grouping = True)
        # Print the results
        print(f" The average budget of {index}-rated movies is ${formatted_value}.")
