==> part3-SQL.ipynb <==
# Project 3 - Part 3 (Core)
* Part 3: Construct and export a MySQL database using your data.

### Specifically, they want the data from the following files included in your database:
 Title Basics:
* Movie ID (tconst)
* Primary Title
* Start Year
* Runtime (in Minutes)
* Genres

Title Ratings:

* Movie ID (tconst)
* Average Movie Rating
* Number of Votes

The TMDB API Results (multiple files):

* Movie ID
* Revenue
* Budget
* Certification (MPAA Rating)
    
You should normalize the tables as best you can before adding them to your new database.

# imports
import pandas as pd

# import title basics csv
title_basics = pd.read_csv("Data/title_basics.csv.gz")

# check
title_basics.info()
title_basics.head()

# import title ratings
title_ratings = pd.read_csv("Data/title_ratings.csv.gz")

# check
title_ratings.info()
title_ratings.head()

# import tmdb data
tmdb_data = pd.read_csv("Data/tmdb_results_combined.csv.gz", low_memory=False)
#"C:\Users\ROSA\Documents\Data\Movie-Profitability\Data\tmdb_results_2000_to_2022.csv.gz"
# check
tmdb_data.info()
tmdb_data.head()



# Normalizing Genres

# split list of genres into separate columns
title_basics['genres_split'] = title_basics['genres'].str.split(',')
title_basics['genres_split'].head()

# create exploded_genres df
exploded_genres = title_basics.explode('genres_split')

# check
exploded_genres.head()

# get list of unique genres from 'genres_split' column
unique_genres = sorted(exploded_genres['genres_split'].unique())

# check
unique_genres

# Create a new title_genres table

# create title_genres df with 'tconst' and 'genres_split'
# as columns
title_genres = exploded_genres[['tconst', 'genres_split']].copy()

# check
title_genres.head()

# Create a genre mapper dictionary to replace string genres with integers

# create genre mapper dictionary (with dict and zip)

# create list of integers, one for each unique genre
genre_ints = range(len(unique_genres))

# create genre mapper: genres to ints
genre_map = dict(zip(unique_genres, genre_ints))

# check
genre_map

# map genre strings to ints in basics in new 'genre_id' column
title_genres['genre_id'] = title_genres['genres_split'].map(genre_map)

# check
title_genres.head()

# drop 'genres_split' column from title_genres
title_genres.drop(columns = ['genres_split'], inplace = True)

# check
title_genres

# Convert the genre map dictionary into a dataframe.


## Manually make dataframe with named cols from the .keyd and .values
genres = pd.DataFrame({'Genre_Name': genre_map.keys(),
                             'Genre_ID' :genre_map.values()})
genres.head()

# Drop unnecessary columns from title_basics table

# check columns in basics
title_basics.columns

# drop unnecessary columns from title_basics 

# desired info: movie id (tconst), primary title, start year, runtime, and genres 


title_basics.drop(columns = ['titleType',
                            'originalTitle',
                            'isAdult',
                            'endYear',
                            'genres',
                            'genres_split'], inplace = True)

# check
title_basics.info()

# Drop unnecessary columns from tmdb_data table

# check columns in tmdb_data
tmdb_data.columns

# check difference between 'imdb_id' and 'id' columns
tmdb_data[['imdb_id', 'id']].head()

# check how many 'imdb_id' == '0'
tmdb_data[tmdb_data['imdb_id'] == '0']

# drop columns with imdb_id as 0
tmdb_data = tmdb_data[tmdb_data['imdb_id'] != '0']

# check
tmdb_data[tmdb_data['imdb_id'] == '0']

# Save tables to MySQL database

# imports
import pymysql
pymysql.install_as_MySQLdb()
from sqlalchemy import create_engine
from sqlalchemy.types import *
from sqlalchemy_utils import database_exists, create_database

# create connection with MySQL
username = 'root'
password = 'Hulk3385smash!'
db_name = 'movies'
connection = f"mysql+pymysql://root:Hulk3385smash!@localhost/{db_name}"

# create engine
engine = create_engine(connection)

# check
engine

# create new database if doesn't exist
if database_exists(connection):
    print(f"The {db_name} database already exists.")
    
else:
    create_database(connection)
   
    print(f"Database {db_name} created.")


# this function adds the table to the db, designates the 
# primary key if given, and outputs the first 5 rows
# of the table for verification
def add_table(table, table_name, primary_key = None):
    
    # get the dtypes of the columns in the table
    dtypes = table.dtypes
    
    # convert columns and dtypes to dictionary for schema
    dtypes_dict = dtypes.to_dict()
    
    # make empty dictionary to build chema
    schema = {}
    
    # iterate through columns in table to build schema dict
    for col in table.columns:
        
        # set schema for ints
        if table[col].dtype == 'int64':
            schema[col] = Integer()
            
        elif table[col].dtype == 'float64':
            schema[col] = Float()
            
        elif table[col].dtype == 'object':
            
            # get max length of the column
            max_len = table[col].fillna('').map(len).max()
            
            # set schema for primary key
            if col == primary_key:
                schema[col] = String(max_len + 1)
                
            else:
                schema[col] = Text(max_len + 1)
    
    # save table to movies db
    table.to_sql(table_name, 
                engine, 
                dtype = schema,
                if_exists = 'replace',
                index = False)
    
    # set primary key column if necessary
    if primary_key != None:
        engine.execute(f'ALTER TABLE {table_name} ADD PRIMARY KEY (`{primary_key}`);')
        
    # check table by printing first five rows
    print(f"First five rows of {table_name} table:")
    
    sql = f"""SELECT *
    FROM {table_name}
    LIMIT 5;
    """
    
    result = pd.read_sql_query(sql, engine)
    print(result)
    
    # check the primary key
    sql = f"""
    SELECT COLUMN_NAME
    FROM INFORMATION_SCHEMA.COLUMNS
    WHERE TABLE_SCHEMA = '{db_name}'
    AND TABLE_NAME = '{table_name}'
    AND COLUMN_KEY = 'PRI';
    """
    result = pd.read_sql_query(sql, engine)
    
    if not result.empty:
        primary_key_column = result.iloc[0]['COLUMN_NAME']
        print(f"The primary key of {table_name} is {primary_key_column}.")
    else:
        print(f"The {table_name} table doesn't have a primary key.")
    
    # print empy lines for readability
    print('\n\n')

# add tables to 'movies' db: title_basics, title_ratings,
# title_genres, genres, and tmdb_data
add_table(title_basics, 'title_basics')
add_table(title_ratings, 'title_ratings', primary_key = 'tconst')
add_table(title_genres, 'title_genres')
add_table(genres, 'genres', primary_key = 'Genre_Name')
add_table(tmdb_data, 'tmdb_data', primary_key = 'imdb_id')

# show tables in movies db
sql = """SHOW TABLES IN movies;
"""

pd.read_sql_query(sql, engine)

# show tables in genres db
q = """
SELECT *
FROM genres
LIMIT 5;
"""

pd.read_sql(q, engine)

# show tables in title_basics db
q = """
SELECT *
FROM title_basics
LIMIT 5;
"""

pd.read_sql(q, engine)

# show tables in title_genres db
q = """
SELECT *
FROM title_genres
LIMIT 5;
"""

pd.read_sql(q, engine)

# show tables in title_ratings db
q = """
SELECT *
FROM title_ratings
LIMIT 5;
"""

pd.read_sql(q, engine)

# show tables in tmdb_data db
q = """
SELECT *
FROM tmdb_data
LIMIT 5;
"""

pd.read_sql(q, engine)

# import os
import os

# specify output file path
output_file = "Data/movies_db.sql"

# export using mysqldump
command = f"mysqldump -u {username} -p{password} -h {'localhost'} {db_name} > {output_file}"

os.system(command)
