==> part4_Hypothesis.ipynb <==
# Hypothesis Testing

### Questions to answer:

* Does the MPAA rating of a movie (G/PG/PG-13/R) affect how much revenue the movie generates?
* Do some movie genres earn more revenue than others?
* Are some genres higher rated than others?

## Deliverables
You should use the same project repository you have been using for Parts 1-3 (for your portfolio).

Create a new notebook in your project repository just for the hypothesis testing (like "Part 4 - Hypothesis Testing.ipynb")

Make sure the results and visualization for all 3 hypotheses are in your notebook.

Please submit the link to your GitHub repository for this assignment.

# imports
import pymysql
pymysql.install_as_MySQLdb()
from sqlalchemy import create_engine
from sqlalchemy.types import *
from sqlalchemy_utils import database_exists
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import scipy.stats as stats
from statsmodels.stats.multicomp import pairwise_tukeyhsd
import statsmodels.stats.multicomp as mc

# create connection with MySQL
username = 'root'
password = 'Hulk3385smash!'
db_name = 'movies'
connection = f"mysql+pymysql://{username}:{password}@localhost/{db_name}"

# create engine
engine = create_engine(connection)

# check
engine

# check by showing tables in db
sql = """SHOW TABLES IN movies;
"""

pd.read_sql_query(sql, engine)

# create helper function to check and remove outliers
# argument 'dictionary' is a dictionary with the groups as keys
# and series of data as values
def check_and_remove_outliers(dictionary):
    
    # iterate over keys (groups) in dictionary
    for key in dictionary.keys():
        
        # check original number of observations
        original_obs = len(dictionary[key])
        
        # check number of outliers
        is_outlier = np.abs(stats.zscore(dictionary[key])) > 3
        number_of_outliers = np.sum(is_outlier)
        
        # remove outliers
        dictionary[key] = dictionary[key][(np.abs(stats.zscore(dictionary[key])) <= 3)]
        
        # print summary
        print(f"Outliers ({number_of_outliers}) removed from group {key};\n",
        f"Number of current observations {len(dictionary[key])} should be {original_obs - number_of_outliers}.")

# Does the MPAA rating of a movie (G/PG/PG-13/R) affect how much revenue the movie generates?

sql = """SELECT t.revenue, t.certification
FROM tmdb_data as t
WHERE t.certification IS NOT NULL
    AND t.certification <> 'NC-17'
    AND t.certification <> 'NR'
    AND t.revenue <> 0;
"""

# save to df
df = pd.read_sql_query(sql, engine)

# check
df.head()

# check that 'certification' only has G, PG, PG-13, and R
df['certification'].apply(lambda x: f"'{x}'").value_counts()
